Suggested Changes to ranks.json
These are my suggested changes for the rank prefixes and name colors.  I feel that the current ones are good, but newcomers might not know what the "❖" rank is and may be more confused. (It could also turn people away). I've also made it so that the "name_color" parameter controls the player's name color (this could be used some way in the future like a perk for changing the name color)